initial visa structs, IpAddress to libnode
