+++ adapter/ph/src/address_pool.rs
@@ -1,10 +1,10 @@
+use libnode::net_defs::IpAddress;
 use openssl::rand::rand_bytes;
 use std::collections::HashSet;
 use std::net::Ipv6Addr;
 use thiserror::Error;
 
 use crate::config;
-use crate::net_defs::IpAddress;
 
 /// Maximum allowed AAA ID value (40 bits)
 ///
