+++ adapter/ph/src/assembly.rs
@@ -12,7 +12,6 @@ use crate::link_state::{LinkEvent, LinkStateError, LinkType};
 use crate::logging::targets::PEER_MGMT;
 use crate::mgmt;
 use crate::mgmt_processor_worker;
-use crate::net_defs::{self, IpAddress, ScopedIpAddr};
 use crate::peer_table;
 use crate::peer_table::PeerInsertError;
 use crate::queues::*;
@@ -24,7 +23,9 @@ use crate::visa_table;
 use crate::vs_types::AuthServicesList;
 use crate::zdp::TerminateReason;
 use crate::zdpr_worker;
+use crate::net_defs::{self, ScopedIpAddr};
 use km_noise::NoiseKeypair;
+use libnode::net_defs::{IpAddress};
 use std::collections::HashMap;
 use std::net::IpAddr;
 use std::num::NonZero;
