+++ adapter/ph/src/mgmt/handlers.rs
@@ -8,12 +8,12 @@ use crate::config;
 use crate::counters;
 use crate::link_state::{LinkEvent, LinkStateError};
 use crate::logging::targets::{FLOW_MGMT, REPORTING, ZDP};
-use crate::net_defs::IpAddress;
 use crate::packet::Packet;
 use crate::tc;
 use crate::tlv::{self, TlvEncoding};
 use crate::zdp;
 use bytes::{Buf, BufMut};
+use libnode::net_defs::IpAddress;
 use std::net::SocketAddr;
 use std::num::NonZero;
 use std::sync::Arc;
