+++ adapter/ph/src/tc.rs
@@ -5,7 +5,7 @@
 //! For now, it's all here together.
 
 use crate::defs;
-use crate::net_defs;
+use libnode::net_defs;
 use crate::zdp;
 use bytes::{Buf, BufMut};
 use zpr_ext::zerocopy::{FromBytesExt, IntoBytesExt};
