+++ adapter/ph/src/visa_mgmt.rs
@@ -2,9 +2,9 @@ use crate::assembly::Assembly;
 use crate::counters::ManagementCounterType;
 use crate::link_state::{LinkEvent, LinkStateError};
 use crate::logging::targets::VISA_MGMT;
-use crate::net_defs::{IPV6_ADDRESS_SIZE, IpAddress};
 use crate::visa_table;
 use crate::vs_types;
+use libnode::net_defs::{IPV6_ADDRESS_SIZE, IpAddress};
 use libnode::{claims, vsapi};
 use std::collections::BTreeMap;
 use std::net::IpAddr;
